Add Windows bat script for training
