+++ run_agent.bat
@@ -1,0 +1,4 @@
+@echo off
+pushd "%~dp0"
+python train.py %*
+popd
